my-requirements-dev.txt
polars-lts-cpu>=1.10.0
pandas
fastexcel
xlsxwriter
pyarrow
pytest-cov
